Applying all requested changes
